Dependencies clean up
