+++ core/build.gradle.kts
@@ -1,25 +1,8 @@
-//import io.gitlab.arturbosch.detekt.Detekt
 import org.jetbrains.kotlin.gradle.plugin.mpp.apple.XCFramework
 
 plugins {
     kotlin("multiplatform") version Versions.kotlin
-    id("maven-publish")
-    id("java-library")
-    id("signing")
-    id("pl.allegro.tech.build.axion-release")/* version Versions.axion*/
-//    id("io.gitlab.arturbosch.detekt") version Versions.detekt
-//    id("io.github.gradle-nexus.publish-plugin") /*version Versions.nexus*/
-    kotlin("plugin.serialization") version Versions.kotlin
     id("io.kotest.multiplatform") version Versions.kotest
-}
-
-apply(from = "../versionConfig.gradle")
-
-group = LibConfig.group
-version = scmVersion.version
-
-val javadocJar = tasks.register("javadocJar", Jar::class.java) {
-    archiveClassifier.set("javadoc")
 }
 
 kotlin {
@@ -80,56 +63,6 @@ kotlin {
         }
     }
 }
-
-publishing {
-    publications.withType<MavenPublication> {
-        artifact(javadocJar)
-
-        pom {
-            name.set(LibConfig.name)
-            description.set("Kotlin multiplatform JsonLogic expressions evaluation engine")
-            url.set(LibConfig.repositoryUrl)
-            inceptionYear.set("2022")
-            licenses {
-                license {
-                    name.set("The Apache License, Version 2.0")
-                    url.set("https://www.apache.org/licenses/LICENSE-2.0.txt")
-                }
-            }
-            developers {
-                developer {
-                    name.set("Marek Krogulski")
-                    email.set("marek.krogulski@allegro.pl")
-                }
-            }
-            scm {
-                connection.set("scm:svn:${LibConfig.repositoryUrl}")
-                developerConnection.set("scm:git@github.com:allegro/json-logic-kmp.git")
-                url.set(LibConfig.repositoryUrl)
-            }
-        }
-    }
-}
-
-//nexusPublishing {
-//    repositories {
-//        sonatype {
-//            username.set(System.getenv("SONATYPE_USERNAME"))
-//            password.set(System.getenv("SONATYPE_PASSWORD"))
-//        }
-//    }
-//}
-
-//System.getenv("GPG_KEY_ID")?.let { gpgKeyId ->
-//    signing {
-//        useInMemoryPgpKeys(
-//            gpgKeyId,
-//            System.getenv("GPG_PRIVATE_KEY"),
-//            System.getenv("GPG_PRIVATE_KEY_PASSWORD")
-//        )
-//        sign(publishing.publications)
-//    }
-//}
 
 /*
 We need to disable this task, because it fails with following error:
